requirements.txt
ag2==0.9.7
aiofiles==24.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.9.0
asyncer==0.0.8
autogen==0.9.7
autogen-agentchat==0.7.5
autogen-core==0.7.5
autogen-ext==0.7.5
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.3.1
colorama==0.4.6
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
fastapi==0.124.4
Flask==3.1.2
groq==0.30.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
jsonref==1.1.0
MarkupSafe==3.0.3
numpy==2.3.5
openai==1.97.1
opentelemetry-api==1.35.0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
protobuf==5.29.5
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pywin32==311
PyYAML==6.0.3
regex==2024.11.6
requests==2.32.4
scikit-learn==1.8.0
scipy==1.16.3
six==1.17.0
sniffio==1.3.1
starlette==0.50.0
termcolor==3.1.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.3
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
Werkzeug==3.1.4
zipp==3.23.0
